+++ tests/lint/float-equality-datetime.pas
@@ -1,0 +1,13 @@
+unit FDT;
+
+interface
+
+implementation
+
+procedure P(D: TDateTime);
+begin
+  if D = 0 then
+    WriteLn('a');
+end;
+
+end.