+++ tests/lint/length-zero-compare.pas
@@ -1,0 +1,15 @@
+unit LZC;
+
+interface
+
+implementation
+
+procedure P(S: string; I: Integer);
+begin
+  if Length(S) = 0 then
+    WriteLn('a');
+  if I = 0 then
+    WriteLn('b');
+end;
+
+end.