+++ tests/lint/outputdebugstring.pas
@@ -1,0 +1,12 @@
+unit ODS;
+
+interface
+
+implementation
+
+procedure P;
+begin
+  OutputDebugString('hi');
+end;
+
+end.